crm_vendas/requirements_supabase.txt
Flask==2.3.3
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
supabase-py==0.0.2
customtkinter==5.2.0
Werkzeug==2.3.7
Flask-Login==0.6.2
python-jose==3.3.0
